django-react/backend/requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
stripe==8.5.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
